Add git rebase alias
Some how `grb` alias is already defined as `git rebase -i HEAD~$1`,
which seems like what I wanted but not.
Also found that unable to override alias with same name and content,
neithor with function with same name and content.
So define another function and use alias `grb` to call that function to
override the original one.