Replace C-p keymap with custom function
Background
==========
Telescope git_files only works in git repo.
If current folder is not a git dir, git_files got nothing.
This behavior is not like fzf, which handles git and non-git dir well in
same keymapping.

Approach
==========
Follow solution in telescope issue, checked if current dir is git or
not, then decide to use git_files or find_files.

ref: https://github.com/nvim-telescope/telescope.nvim/wiki/Configuration-Recipes#falling-back-to-find_files-if-git_files-cant-find-a-git-directory