Create ResolveDbConflict command
Reason
==========
It's troublesome when trying to resolve merge conflicts in db/schema.rb
and db/structure.sql, better make it as a command to save time.

The function is created by chatGPT, below are details

Model
==========
GPT-4o

Initial Prompt
==============
you're an expert of nvim, able to configure nvim or write functions in lua.
Write a function that able to solve the selected merge conflict in db schema version as below.
There'll be two kinds of conflict pattern in different files.

a. solve conflict in `db/schema.rb`.
Below is the example when having conflict.
You need to pick the latest version of schema, which is `2025_06_11_004922` in below example.
  ```
  # before
  <<<<<<< HEAD
  ActiveRecord::Schema.define(version: 2025_06_10_062940) do
  ||||||| fa1a4179d2
  ActiveRecord::Schema.define(version: 2025_06_03_064951) do
  =======
  ActiveRecord::Schema.define(version: 2025_06_11_004922) do
  >>>>>>> master

  # after
  ActiveRecord::Schema.define(version: 2025_06_11_004922) do
  ```

b. solve conflict in `db/structure.sql`.
Below is the example when having conflict.
You need to
1. sort the schema migration versions `('xxx')`
2. remove duplicated schema migration versions
3. remove conflict indicators like `<<<<<<< HEAD`, `||||||| fa1a4179d2`, `=======` and `>>>>>>> master`
4. make sure the most bottom version end with `;`, such as `('20250611004922');`. And other versions should end with `,`
  ```
  # before
  <<<<<<< HEAD
  ('20250603042308'),
  ('20250603064951'),
  ('20250610062940');

  ||||||| fa1a4179d2
  ('20250603064951');
  ('20250603042308');
  =======
  ('20250603042308'),
  ('20250603064951'),
  ('20250605094539'),
  ('20250611004922');

  >>>>>>> master

  # after
  ('20250603042308'),
  ('20250603064951'),
  ('20250605094539'),
  ('20250610062940'),
  ('20250611004922');
  ```

In the example code block above, `# before` is the original code when having conflict, `# after` is
the expected result.

Refinement Prompts
==================
- I don't need to apply range, please remove it.
- found  two issues
  1. got `No range allowed` when trying to run the command in select
     mode
  2. I run the command in normal mode, found the filename not matched
     the pattern. The filename is a full path like
     `/proj/nerv_ave_ck/db/schema.rb` that doesn't begin with `db/` .